requirements.txt
streamlit
pandas
numpy
librosa
scikit-learn
joblib
tqdm
streamlit-audiorecorder